Refine and refactor the proof and file structure
Compact some similar lemmas. Try to delete some useless functions. Move
some lemmas to common module (i.e, kubernetes_cluster::proof, not specific 
to one controller). Refactor the file structure to make it more reasonable.
Reword some comments and naming. Try to make the verification more stable.
Rename lower_rv => smaller_rv.

---------
