+++ src/controller_examples/rabbitmq_controller/proof/helper_invariants/mod.rs
