+++ src/controller_examples/rabbitmq_controller/proof/safety/mod.rs
@@ -1,4 +1,3 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-pub mod liveness;
-pub mod terminate;
+pub mod safety;
