+++ src/kubernetes_cluster/proof/mod.rs
@@ -11,3 +11,4 @@ pub mod kubernetes_api_liveness;
 pub mod kubernetes_api_safety;
 pub mod message;
 pub mod wf1_assistant;
+pub mod validation_rule;
